Add Standalone mode to share single dashboard among multiple services (#95)
* imp: standalone mode

* refactor: improve standalone mode handling and logging

* refactor: logging for standalone mode

---------

+++ src/AspireRunner.Core/Dashboard.cs
@@ -1,4 +1,4 @@
-﻿using AspireRunner.Core.Extensions;
+using AspireRunner.Core.Extensions;
 using Medallion.Threading.FileSystem;
 using Microsoft.Extensions.Logging;
 using System.Diagnostics;
@@ -83,9 +83,16 @@ public partial class Dashboard : IDashboard
             return;
         }
 
+        _stopRequested = true;
+        if (Options.Runner.Mode is RunningMode.Standalone)
+        {
+            _dashboardProcess = null;
+            LogStopIgnoredStandaloneMode();
+            return;
+        }
+
         try
         {
-            _stopRequested = true;
             await Task.Run(() => _dashboardProcess?.Kill(true), cancellationToken);
         }
         catch (InvalidOperationException)
@@ -115,14 +122,28 @@ public partial class Dashboard : IDashboard
 
             if (instance.Dashboard.IsRunning())
             {
-                if (Options.Runner.SingleInstanceHandling is SingleInstanceHandling.ReplaceExisting || !instance.Runner.IsRunning())
+                if (Options.Runner.Mode is RunningMode.Standalone)
+                {
+                    // Reuse the existing dashboard process instead of restarting
+                    _dashboardProcess = instance.Dashboard;
+                    LogReusingRunningDashboard(_dashboardProcess.Id);
+
+                    if (Options.Runner.RestartOnFailure)
+                    {
+                        RegisterProcessExitHandler();
+                    }
+
+                    return true;
+                }
+
+                if (!instance.Runner.IsRunning() || Options.Runner.SingleInstanceHandling is SingleInstanceHandling.ReplaceExisting)
                 {
                     instance.Dashboard.Kill(true);
                 }
                 else if (Options.Runner.SingleInstanceHandling is SingleInstanceHandling.WarnAndExit)
                 {
                     WarnExistingInstance(instance.Dashboard.Id);
-                    return false;
+                    return true;
                 }
             }
